+++ Assets/Plugins/iOS/EOSManager_iOS.cs
@@ -33,7 +33,7 @@ using Epic.OnlineServices.Auth;
 using Epic.OnlineServices.Logging;
 using System.Runtime.InteropServices;
 
-#if UNITY_IOS && !UNITY_EDITOR
+#if UNITY_IOS && !UNITY_EDITOR && EOS_PREVIEW_PLATFORM
 namespace PlayEveryWare.EpicOnlineServices
 {
     //-------------------------------------------------------------------------
@@ -205,6 +205,55 @@ namespace PlayEveryWare.EpicOnlineServices
         }
 
         //-------------------------------------------------------------------------
+        /// <summary>
+        /// Create App Controller necessary for iOS login options
+        /// </summary>
+        [DllImport("__Internal")]
+        static private extern IntPtr LoginUtility_get_app_controller();
+
+        //-------------------------------------------------------------------------
+        /// <summary>
+        /// Make Login Options for iOS Specific
+        /// </summary>
+        public IOSLoginOptions MakeIOSLoginOptionsFromDefualt(Epic.OnlineServices.Auth.LoginOptions loginOptions)
+        {
+            IOSLoginOptions modifiedLoginOptions = new IOSLoginOptions();
+            modifiedLoginOptions.ScopeFlags = loginOptions.ScopeFlags;
+
+            var credentials = new IOSCredentials();
+
+            credentials.Token = loginOptions.Credentials.Value.Token;
+            credentials.Id = loginOptions.Credentials.Value.Id;
+            credentials.Type = loginOptions.Credentials.Value.Type;
+            credentials.ExternalType = loginOptions.Credentials.Value.ExternalType;
+
+            var systemAuthCredentialsOptions = new IOSCredentialsSystemAuthCredentialsOptions();
+
+            systemAuthCredentialsOptions.PresentationContextProviding = LoginUtility_get_app_controller();
+            credentials.SystemAuthCredentialsOptions = systemAuthCredentialsOptions;
+
+            modifiedLoginOptions.Credentials = credentials;
+
+            return modifiedLoginOptions;
+        }
+
+        //-------------------------------------------------------------------------
+        /// <summary>
+        /// Set Default Audio Session for iOS (Category to AVAudioSessionCategoryPlayAndRecord)
+        /// </summary>
+        [DllImport("__Internal")]
+        static private extern void MicrophoneUtility_set_default_audio_session();
+
+        //-------------------------------------------------------------------------
+        /// <summary>
+        /// Set Default Audio Session for iOS
+        /// </summary>
+        public void SetDefaultAudioSession()
+        {
+            MicrophoneUtility_set_default_audio_session();
+        }
+        //-------------------------------------------------------------------------
+
         public void InitializeOverlay(IEOSCoroutineOwner owner)
         {
         }